Define new /build APIs
