+++ packages/server/src/service/model.ts
@@ -1,4 +1,5 @@
 import { DuckDBConnection } from "@malloydata/db-duckdb";
+import type { BuildManifest } from "@malloydata/malloy";
 import {
    Annotation,
    API,
@@ -125,17 +126,11 @@ export class Model {
       packagePath: string,
       modelPath: string,
       connections: Map<string, Connection>,
-      options?: { buildManifest?: Record<string, { tableName: string }> },
    ): Promise<Model> {
       // getModelRuntime might throw a ModelNotFoundError. It's the callers responsibility
       // to pass a valid model path or handle the error.
       const { runtime, modelURL, importBaseURL, dataStyles, modelType } =
-         await Model.getModelRuntime(
-            packagePath,
-            modelPath,
-            connections,
-            options,
-         );
+         await Model.getModelRuntime(packagePath, modelPath, connections);
 
       try {
          const { modelMaterializer, runnableNotebookCells } =
@@ -260,7 +255,7 @@ export class Model {
    }
 
    public getQueries(): ApiQuery[] | undefined {
-      return this.sources;
+      return this.queries;
    }
 
    public async getModel(): Promise<ApiCompiledModel> {
@@ -298,6 +293,7 @@ export class Model {
       sourceName?: string,
       queryName?: string,
       query?: string,
+      options?: { buildManifest?: BuildManifest },
    ): Promise<{
       result: Malloy.Result;
       compactResult: QueryData;
@@ -323,12 +319,19 @@ export class Model {
          throw new BadRequestError("Model has no queryable entities.");
 
       // Wrap loadQuery calls in try-catch to handle query parsing errors
+      const queryOptions = options?.buildManifest
+         ? { buildManifest: options.buildManifest }
+         : undefined;
       try {
          if (!sourceName && !queryName && query) {
-            runnable = this.modelMaterializer.loadQuery("\n" + query);
+            runnable = this.modelMaterializer.loadQuery(
+               "\n" + query,
+               queryOptions,
+            );
          } else if (queryName && !query) {
             runnable = this.modelMaterializer.loadQuery(
                `\nrun: ${sourceName ? sourceName + "->" : ""}${queryName}`,
+               queryOptions,
             );
          } else {
             const endTime = performance.now();
@@ -576,7 +579,6 @@ export class Model {
       packagePath: string,
       modelPath: string,
       connections: Map<string, Connection>,
-      options?: { buildManifest?: Record<string, { tableName: string }> },
    ): Promise<{
       runtime: Runtime;
       modelURL: URL;
@@ -616,26 +618,10 @@ export class Model {
          `SET FILE_SEARCH_PATH='${workingDirectory}';`,
       );
 
-      const runtimeOptions: {
-         urlReader: typeof urlReader;
-         connections: FixedConnectionMap;
-         buildManifest?: {
-            entries: Record<string, { tableName: string }>;
-            strict?: boolean;
-         };
-      } = {
+      const runtime = new Runtime({
          urlReader,
          connections: new FixedConnectionMap(connections, "duckdb"),
-      };
-
-      if (options?.buildManifest) {
-         runtimeOptions.buildManifest = {
-            entries: options.buildManifest,
-            strict: false,
-         };
-      }
-
-      const runtime = new Runtime(runtimeOptions);
+      });
       const dataStyles = urlReader.getHackyAccumulatedDataStyles();
       return { runtime, modelURL, importBaseURL, dataStyles, modelType };
    }
