+++ packages/server/src/storage/duckdb/DuckDBRepository.ts
@@ -1,32 +1,31 @@
 import {
+   BuildExecution,
+   BuildExecutionStatus,
    Connection,
    ManifestEntry,
    Package,
    Project,
    ResourceRepository,
-   Task,
-   TaskExecution,
-   TaskExecutionStatus,
 } from "../DatabaseInterface";
+import { BuildExecutionRepository } from "./BuildExecutionRepository";
 import { ConnectionRepository } from "./ConnectionRepository";
 import { DuckDBConnection } from "./DuckDBConnection";
 import { ManifestRepository } from "./ManifestRepository";
 import { PackageRepository } from "./PackageRepository";
 import { ProjectRepository } from "./ProjectRepository";
-import { TaskRepository } from "./TaskRepository";
 
 export class DuckDBRepository implements ResourceRepository {
    private projectRepo: ProjectRepository;
    private packageRepo: PackageRepository;
    private connectionRepo: ConnectionRepository;
-   private taskRepo: TaskRepository;
+   private buildExecRepo: BuildExecutionRepository;
    private manifestRepo: ManifestRepository;
 
    constructor(public db: DuckDBConnection) {
       this.projectRepo = new ProjectRepository(db);
       this.packageRepo = new PackageRepository(db);
       this.connectionRepo = new ConnectionRepository(db);
-      this.taskRepo = new TaskRepository(db);
+      this.buildExecRepo = new BuildExecutionRepository(db);
       this.manifestRepo = new ManifestRepository(db);
    }
 
@@ -59,7 +58,7 @@ export class DuckDBRepository implements ResourceRepository {
 
    async deleteProject(id: string): Promise<void> {
       await this.manifestRepo.deleteEntriesByProjectId(id);
-      await this.taskRepo.deleteTasksByProjectId(id);
+      await this.buildExecRepo.deleteByProjectId(id);
       await this.connectionRepo.deleteConnectionsByProjectId(id);
       await this.packageRepo.deletePackagesByProjectId(id);
       await this.projectRepo.deleteProject(id);
@@ -96,7 +95,15 @@ export class DuckDBRepository implements ResourceRepository {
    }
 
    async deletePackage(id: string): Promise<void> {
-      return this.packageRepo.deletePackage(id);
+      const pkg = await this.packageRepo.getPackageById(id);
+      if (pkg) {
+         await this.manifestRepo.deleteEntriesByPackage(
+            pkg.projectId,
+            pkg.name,
+         );
+         await this.buildExecRepo.deleteByPackage(pkg.projectId, pkg.name);
+      }
+      await this.packageRepo.deletePackage(id);
    }
 
    async deletePackagesByProjectId(id: string): Promise<void> {
@@ -141,70 +148,45 @@ export class DuckDBRepository implements ResourceRepository {
       return this.connectionRepo.deleteConnectionsByProjectId(id);
    }
 
-   // ==================== TASKS ====================
+   // ==================== BUILD EXECUTIONS ====================
 
-   async listTasks(projectId: string): Promise<Task[]> {
-      return this.taskRepo.listTasks(projectId);
+   async listBuildExecutions(
+      projectId: string,
+      packageName: string,
+   ): Promise<BuildExecution[]> {
+      return this.buildExecRepo.listExecutions(projectId, packageName);
    }
 
-   async getTaskById(id: string): Promise<Task | null> {
-      return this.taskRepo.getTaskById(id);
+   async getBuildExecutionById(id: string): Promise<BuildExecution | null> {
+      return this.buildExecRepo.getExecutionById(id);
    }
 
-   async getTaskByName(projectId: string, name: string): Promise<Task | null> {
-      return this.taskRepo.getTaskByName(projectId, name);
+   async getRunningBuildExecution(
+      projectId: string,
+      packageName: string,
+   ): Promise<BuildExecution | null> {
+      return this.buildExecRepo.getRunningExecution(projectId, packageName);
    }
 
-   async createTask(
-      task: Omit<Task, "id" | "createdAt" | "updatedAt">,
-   ): Promise<Task> {
-      return this.taskRepo.createTask(task);
+   async createBuildExecution(
+      projectId: string,
+      packageName: string,
+      status: BuildExecutionStatus = "PENDING",
+   ): Promise<BuildExecution | null> {
+      return this.buildExecRepo.createExecution(projectId, packageName, status);
    }
 
-   async updateTask(id: string, updates: Partial<Task>): Promise<Task> {
-      return this.taskRepo.updateTask(id, updates);
-   }
-
-   async deleteTask(id: string): Promise<void> {
-      return this.taskRepo.deleteTask(id);
-   }
-
-   async deleteTasksByProjectId(projectId: string): Promise<void> {
-      return this.taskRepo.deleteTasksByProjectId(projectId);
-   }
-
-   // ==================== TASK EXECUTIONS ====================
-
-   async listExecutions(taskId: string): Promise<TaskExecution[]> {
-      return this.taskRepo.listExecutions(taskId);
-   }
-
-   async getExecutionById(id: string): Promise<TaskExecution | null> {
-      return this.taskRepo.getExecutionById(id);
-   }
-
-   async getRunningExecution(taskId: string): Promise<TaskExecution | null> {
-      return this.taskRepo.getRunningExecution(taskId);
-   }
-
-   async createExecution(
-      taskId: string,
-      status: TaskExecutionStatus = "PENDING",
-   ): Promise<TaskExecution | null> {
-      return this.taskRepo.createExecution(taskId, status);
-   }
-
-   async updateExecution(
+   async updateBuildExecution(
       id: string,
       updates: {
-         status?: TaskExecutionStatus;
+         status?: BuildExecutionStatus;
          startedAt?: Date;
          completedAt?: Date;
          error?: string | null;
          metadata?: Record<string, unknown> | null;
       },
-   ): Promise<TaskExecution> {
-      return this.taskRepo.updateExecution(id, updates);
+   ): Promise<BuildExecution> {
+      return this.buildExecRepo.updateExecution(id, updates);
    }
 
    // ==================== BUILD MANIFESTS ====================
@@ -216,30 +198,6 @@ export class DuckDBRepository implements ResourceRepository {
       return this.manifestRepo.listEntries(projectId, packageName);
    }
 
-   async getManifestEntryByBuildId(
-      projectId: string,
-      packageName: string,
-      buildId: string,
-   ): Promise<ManifestEntry | null> {
-      return this.manifestRepo.getEntryByBuildId(
-         projectId,
-         packageName,
-         buildId,
-      );
-   }
-
-   async getManifestEntryBySourceName(
-      projectId: string,
-      packageName: string,
-      sourceName: string,
-   ): Promise<ManifestEntry | null> {
-      return this.manifestRepo.getEntryBySourceName(
-         projectId,
-         packageName,
-         sourceName,
-      );
-   }
-
    async upsertManifestEntry(
       entry: Omit<ManifestEntry, "id" | "createdAt" | "updatedAt">,
    ): Promise<ManifestEntry> {
@@ -248,16 +206,5 @@ export class DuckDBRepository implements ResourceRepository {
 
    async deleteManifestEntry(id: string): Promise<void> {
       return this.manifestRepo.deleteEntry(id);
-   }
-
-   async deleteManifestEntriesByPackage(
-      projectId: string,
-      packageName: string,
-   ): Promise<void> {
-      return this.manifestRepo.deleteEntriesByPackage(projectId, packageName);
-   }
-
-   async deleteManifestEntriesByProjectId(projectId: string): Promise<void> {
-      return this.manifestRepo.deleteEntriesByProjectId(projectId);
    }
 }
